refactor script set tableau
